requirements.txt
Flask==0.12
Flask-SQLAlchemy==2.1
Flask-Migrate==2.0.3
Flask-Admin==1.4.2
Flask-Cors==3.0.2
Flask-JWT==0.3.2
Flask-Script==2.0.5
Flask-Mail==0.9.1
marshmallow==2.12.2
redis==2.10.5
celery==4.0.2
requests==2.13.0
psycopg2==2.6.2
boto3==1.4.4
flake8==3.3.0
ipython==5.2.2
konch==2.3.0
Flask-WebTest==0.0.8
flask-pytest==0.0.5
